gcc-cross-wrapper: Remove in favor of newly-cross-supporting `cc-wrapper`
Besides the support for prefixed binary names in the last commit, we
also support per-target flag interposition by putting the triple in the
names of the relevant environment variables, e.g:

export NIX_arm_unknown_linux_gnu_CFLAGS_COMPILE=...

Note how that the dashes are replaced to keep the identifier valid.
Using names like this allows us to keep the settings for different
compilers seperate.

I think it might be even better to use names like `NIX_{BUILD,HOST}...`
using the platform's role rather than the platform itself, but this
would be more work as the previous stages' tools would have to be re-
wrapped to take on their new role. I therefore didn't do this for now,
but that route should be thoroughly explored in the future.

 - glibc built in right stage
 - no more gcc-cross-wrapper, just cc-wrapper
 - no more ccCross binutilsCross in stdenv, stdenv has right gcc and
   only that
 - linuxHeadersCross also purged, just select via stage

# Conflicts:
#	pkgs/top-level/all-packages.nix